Test utils: remove the ability to clear the registry.
Instead we provide the ability to "freeze" the current registry and
diff against its previous state. This makes tests much easier to write
as different tests can now influence the same metrics safely.

This requires prometheus_client >= 0.0.13. I believe that at this
point we are not using private APIs of the prometheus client anymore.