Fix the import of MigrationExecutor to be lazy.
MigrationExecutor checks at import time whether apps are all
loaded. This doesn't work as they are usually not if your app is
imported by adding it to INSTALLED_APPS. Importing the MigrationExecutor
locally solves this problem as ExportMigration is only called once the
django_prometheus app is signaled that it's ready.