Celery: collect per-task metrics using the push gateway
Since Celery runs tasks in multiple worker processes it's
not possible to expose the metrics directly since there's
no guarantee that a worker process will still be running by
the time the next scrape runs and we wouldn't want to have
to scrape dozens of possible endpoints anyway.

The included signal handlers will push task start and stop
metrics when a push gateway is defined in your settings.